requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
Pillow==11.0.0
yookassa==3.4.0
certifi==2024.8.30
asyncpg==0.30.0
